organized classes in respective files, added more signals to View
